Parallelization
Reimplemented the `create_app` function to create processes for
individual source/filter/output modules instead of running the modules
directly in a sequential loop in the main method.

Instead of this the main app loop simply follows a sequence of
multiprocessing.connection.Connection calls to send and receive data
from the parallel processes.

In-between the process communication we do additional "merge" logic
similar to what we've done previously as this likely helps with
improving some efficiency.

This should improve the overall performance of the tool at the expense
of burning more cpu cores for the parallel processes.

Signed-off-by: Ondrej Lichtner <ilichtner@fit.vutbr.cz>